Fix compiler issues

+++ BinaryIO.cpp
@@ -8,10 +8,10 @@ void bio::writeString(std::vector<uint8_t>& data, const std::string& val) {
 	writeSimple<char>(data, '\0');
 }
 
-std::string bio::readString(const std::vector<uint8_t>& data, int& at) {
+std::string bio::readString(const std::vector<uint8_t>& data, size_t& at) {
 	std::string s = "";
 	assert(at < data.size());
-	while (at < (int)data.size()) {
+	while (at < data.size()) {
 		char c = readSimple<char>(data, at);
 		if (c == '\0') break;
 		else s += c;
